Refactor to use CharacterDetails struct in snapshots
