+++ RickAndMorty-iOS/Views/Shared/Section.swift
@@ -11,7 +11,7 @@ enum Section: Int, CaseIterable {
     case appearance
     case info
     case location
-    case characters
+    case episodes
     var columnCount: Int {
         switch self {
         case .appearance:
@@ -20,7 +20,7 @@ enum Section: Int, CaseIterable {
             return 3
         case .location:
             return 2
-        case .characters:
+        case .episodes:
             return 2
         }
     }
